+++ lib/src/in_app_browser/in_app_browser_settings.dart
@@ -60,11 +60,6 @@ class InAppBrowserClassSettings {
     return options;
   }
 
-  static Map<String, dynamic> instanceToMap(
-      InAppBrowserClassSettings settings) {
-    return settings.toMap();
-  }
-
   Map<String, dynamic> toJson() {
     return this.toMap();
   }