+++ lib/src/types/main.dart
@@ -145,3 +145,5 @@ export 'window_features.dart';
 export 'web_resource_error.dart';
 export 'web_resource_error_type.dart';
 export 'media_capture_state.dart';
+export 'proxy_rule.dart';
+export 'proxy_scheme_filter.dart';