+++ nodejs_server_test_auth_basic_and_ssl/index.js
@@ -1,12 +1,14 @@
 // Example of the server https is taken from here: https://engineering.circle.com/https-authorized-certs-with-node-js-315e548354a2
 // Conversion of client1-crt.pem to certificate.pfx: https://stackoverflow.com/a/38408666/4637638
-const express = require('express')
-const https = require('https')
-const cors = require('cors')
-const auth = require('basic-auth')
-const app = express()
-const appHttps = express()
-const appAuthBasic = express()
+const express = require('express');
+const proxy = require('express-http-proxy');
+const https = require('https');
+const cors = require('cors');
+const auth = require('basic-auth');
+const app = express();
+const appHttps = express();
+const appAuthBasic = express();
+const appProxy = express();
 const fs = require('fs')
 const path = require('path')
 const bodyParser = require('body-parser');
@@ -20,6 +22,7 @@ var options = {
   rejectUnauthorized: false
 };
 
+appHttps.use('/', proxy('www.google.com'));
 appHttps.get('/', (req, res) => {
   console.log(JSON.stringify(req.headers))
 	const cert = req.connection.getPeerCertificate()
@@ -197,3 +200,7 @@ app.get("/test-download-file", (req, res) => {
 })
 
 app.listen(8082)
+
+appProxy.use(cors());
+appProxy.use('/', proxy('www.google.com'));
+appProxy.listen(8083);